feat: Notifications: Refactor NotificationManagerWrapper (SQCORE-1199) (#3624)
* Remove NotificationData id - not used anymore

* Splitting NotificationManagerWrapper.scala into a number of smaller files

* fix: Ensure notifications are being cancelled (SQCORE-1199) (#3626)

* fix: Ensure notifications are being cancelled (SQCORE-1199)

* feat: Show notifications when the app is open but on a different conversation (SQCORE-1197)

* feat: Make sure the new system works with  Federation (SQCORE-1181)

* fix: More aggressive cancelling of notifications (SQCORE-1199)

* fix: Notice changes in the availability status
