+++ app/src/main/scala/com/waz/services/notifications/NotificationsHandlerService.scala
@@ -27,14 +27,12 @@ import com.waz.services.FutureService
 import com.waz.threading.Threading
 import com.waz.utils.{TimedWakeLock, returning}
 import com.waz.zclient.ServiceHelper
-import com.waz.zclient.log.LogUI._
 import com.waz.zclient.notifications.controllers.NotificationManagerWrapper
 
 import scala.concurrent.Future
 import scala.concurrent.duration._
 
 class NotificationsHandlerService extends FutureService with ServiceHelper with DerivedLogTag {
-
   import NotificationsHandlerService._
   import Threading.Implicits.Background
 
@@ -43,16 +41,14 @@ class NotificationsHandlerService extends FutureService with ServiceHelper with 
   private lazy val notificationManager = inject[NotificationManagerWrapper]
 
   override protected def onIntent(intent: Intent, id: Int): Future[Any] = wakeLock.async {
-
-    val account = Option(intent.getStringExtra(ExtraAccountId)).map(UserId(_))
-    val conversation = Option(intent.getStringExtra(ExtraConvId)).map(ConvId(_))
-    val instantReplyContent = Option(RemoteInput.getResultsFromIntent(intent)).map(_.getCharSequence(InstantReplyKey))
+    def account = Option(intent.getStringExtra(ExtraAccountId)).map(UserId(_))
+    def conversation = Option(intent.getStringExtra(ExtraConvId)).map(ConvId(_))
+    def instantReplyContent = Option(RemoteInput.getResultsFromIntent(intent)).map(_.getCharSequence(InstantReplyKey))
 
     Option(ZMessaging.currentAccounts) match {
       case Some(accs) =>
         (account, conversation, instantReplyContent) match {
           case (Some(acc), Some(convId), _) if ActionClear == intent.getAction =>
-            verbose(l"Clearing notifications for account: $acc and conversation:$conversation")
             Future.successful(notificationManager.cancelNotifications(acc, Set(convId)))
           case (Some(acc), Some(convId), Some(content)) if ActionQuickReply == intent.getAction =>
             accs.getZms(acc).flatMap {
@@ -61,15 +57,13 @@ class NotificationsHandlerService extends FutureService with ServiceHelper with 
                   notificationManager.cancelNotifications(acc, Set(convId))
                 }
               case _ =>
-                Future.successful({})
+                Future.successful(())
             }
           case _ =>
-            warn(l"No account id passed on intent")
-            Future.successful({})
+            Future.successful(())
         }
       case _ =>
-        warn(l"No AccountsService available")
-        Future.successful({})
+        Future.successful(())
     }
   }
 }
