+++ app/src/main/scala/com/waz/zclient/notifications/controllers/SpannableWrapper.scala
@@ -1,0 +1,64 @@
+package com.waz.zclient.notifications.controllers
+
+import android.content.Context
+import android.graphics.{Color, Typeface}
+import android.text.{SpannableString, Spanned}
+import android.text.style.{ForegroundColorSpan, StyleSpan}
+import com.waz.utils.returning
+import com.waz.zclient.utils.{ResString, format}
+
+final case class SpannableWrapper(header: ResString,
+                                  body: ResString,
+                                  spans: List[Span],
+                                  separator: String) {
+  override def toString: String =
+    format(className = "SpannableWrapper", oneLiner = true,
+      "header"    -> Some(header),
+      "body"      -> Some(body),
+      "spans"     -> (if (spans.nonEmpty) Some(spans) else None),
+      "separator" -> (if (separator.nonEmpty) Some(separator) else None)
+    )
+
+  def build(implicit cxt: Context): SpannableString = {
+    val headerStr = header.resolve
+    val bodyStr = body.resolve
+    val wholeStr = headerStr + separator + bodyStr
+
+    def range(span: Span) = span.range match {
+      case Span.HeaderRange => (0, headerStr.length)
+      case Span.BodyRange   => (headerStr.length + span.offset, wholeStr.length)
+    }
+
+    def style(span: Span) = span.style match {
+      case Span.ForegroundColorSpanBlack => new ForegroundColorSpan(Color.BLACK)
+      case Span.ForegroundColorSpanGray  => new ForegroundColorSpan(Color.GRAY)
+      case Span.StyleSpanBold            => new StyleSpan(Typeface.BOLD)
+      case Span.StyleSpanItalic          => new StyleSpan(Typeface.ITALIC)
+    }
+
+    returning(new SpannableString(wholeStr)) { ss =>
+      spans.map(span => (style(span), range(span))).foreach {
+        case (style, (start, end)) if end > start => ss.setSpan(style, start, end, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE)
+        case _ =>
+      }
+    }
+  }
+
+  def +(span: Span): SpannableWrapper = copy(spans = this.spans ++ List(span))
+
+  def +(sw: SpannableWrapper): SpannableWrapper = {
+    val spans     = this.spans ++ sw.spans.map(span => if (span.range == Span.HeaderRange) span.copy(range = Span.BodyRange) else span)
+    val body      = if (sw.header != ResString.Empty) sw.header else sw.body
+    copy(spans = spans, body = body)
+  }
+}
+
+object SpannableWrapper {
+  def apply(header: ResString): SpannableWrapper =
+    SpannableWrapper(header = header, body = ResString.Empty, spans = List.empty, separator = "")
+  def apply(header: ResString, spans: List[Span]): SpannableWrapper =
+    SpannableWrapper(header = header, body = ResString.Empty, spans = spans, separator = "")
+
+  val Empty: SpannableWrapper = SpannableWrapper(ResString.Empty)
+}
+
